+++ src/Views/Controls/ContentTabItemView.xaml.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Security.Policy;
+using System.Text;
+using System.Windows;
+using CefSharp.Wpf;
+
+namespace EpubViewer
+{
+    public partial class ContentTabItemView:ChromiumWebBrowser
+    {
+    }
+}